Feature/dcc 490  npe de ped brain

Former-commit-id: 4291bdf40a7fdf3b74afa282798e674d1e5cf35b
Former-commit-id: 69a0b92d43f456db754755ab23a0481e11fde37a

+++ server/src/main/java/org/icgc/dcc/validation/report/ValidationErrorReport.java
@@ -114,7 +114,7 @@ public class ValidationErrorReport {
     for(ColumnErrorReport column : this.columns) {
       offsets.addAll(column.getLines());
     }
-    Map<Long, Integer> byteToLine = ByteOffsetToLineNumber.convert(file, offsets);
+    Map<Long, Long> byteToLine = ByteOffsetToLineNumber.convert(file, offsets);
 
     if(byteToLine != null) {
       for(ColumnErrorReport column : this.columns) {
